layer-shell: Update cursor when arranging surfaces
The code so far only updated the cursor for the layer surface
but we need to take the case into account where a layer surface
moves away and reveals a non layer-surface.

Otherwise we might get the wrong cursor when we e.g. close
phosh's overview and the cursor is over an app.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/712>