render: Exit early when there is no damage
As touch-points and debug-damage submit proper damage nowadays there's
no point in proceeding. Software cursors are only rendered when they
overlap with the pass in damage either.

This has the nice side effect that we don't emit any tracemarks when
there was nothing to render.

We keep emitting `render-end` for the moment until we received that signal
from the output shields.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/718>