output: Add debug option to fake builtin displays
For nested development it is useful to make an output behave
like a built-in one. Add a debug flag similar to what we did
for phosh a while back. Due to our current test setup we
assume all nested displays are built in, not just the first.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/738>