treewide: Update more copyright
Update copyright for files that only FLA holders modified
in the time frame. If in doubt leave it as is.

We also fix the copyright of idle-inhibit.h and view-deco.h headers
which were copy/paste error on my end as they have the same copyright
than the corresponding C files.

Gbp-Dch: Ignore
Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/733>